+++ src/Point.h
@@ -5,8 +5,8 @@
 
 typedef struct
 {
-    int64_t x;
-    int64_t y;
+    int32_t x;
+    int32_t y;
 }
 Point;
 
@@ -18,9 +18,9 @@ Point Point_Add(const Point a, const Point b);
 
 Point Point_Sub(const Point a, const Point b);
 
-Point Point_Div(const Point a, const int64_t n);
+Point Point_Div(const Point a, const int32_t n);
 
-Point Point_Mul(const Point a, const int64_t n);
+Point Point_Mul(const Point a, const int32_t n);
 
 Point Point_ToIso(const Point);
 
@@ -28,9 +28,9 @@ Point Point_ToCart(const Point);
 
 bool Point_IsZero(const Point);
 
-int64_t Point_Mag(const Point);
+int32_t Point_Mag(const Point);
 
-Point Point_Normalize(const Point, const int64_t normal);
+Point Point_Normalize(const Point, const int32_t normal);
 
 Point Point_Dot(const Point a, const Point b);
